GA_OOS/requirements.txt
apiclient==1.0.3
cachetools==3.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.1
numpy==1.16.2
oauth2client==4.1.3
pandas==0.24.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
rsa==4.0
six==1.12.0
suds-jurko==0.6
uritemplate==3.0.0
urllib3==1.24.1
